requirements.txt
dotenv~=0.9.9
google-genai~=1.24.0
python-dotenv~=1.1.1
numpy
psycopg2